Fix `when/splice' inconsistent return types
**CURRENT BEHAVIOR**

The function `when/splice` returns either a list (when the condition is true), or a string (when the condition is false). This makes it unusable in contexts that expect lists—for example, `(... ,@◊when/splice[...]{...} ...)`.

**EXPECTED BEHAVIOR**

The return types of `when/splice` are consistent. When the condition is false, return an empty list.